refactor: split pure helpers into the core, isolate fs/path in the server layer
The core transformers imported ~/utils/common, whose top level pulls in
fs/path via the image download helper — semantically unused by the
transform but present in its module graph, so it would bundle (and fail)
under platform:neutral. Pure helpers now live in src/core/utils.ts;
downloadFigmaImage keeps the builtins in the server layer.

Also moves the NodeSnapshot type guards (utils/identity) into the core —
they are core domain logic — dropping the unused remeda re-export, which
was remeda's only import site, so the dep goes too. Deletes dead
removeEmptyKeys.
