+++ src/core/transformers/layout/grid.ts
@@ -1,4 +1,4 @@
-import { hasGridLayout, isLayout } from "~/utils/identity.js";
+import { hasGridLayout, isLayout } from "~/core/identity.js";
 import type { NodeSnapshot } from "~/core/snapshot.js";
 import type { SimplifiedLayout } from "./common.js";
