+++ src/core/utils.ts
@@ -1,0 +1,112 @@
+/**
+ * Builtin-free helpers for the canonicalize core. This module must stay pure —
+ * no Node builtins, no external packages — because the whole core graph bundles
+ * into the plugin's QuickJS sandbox (Invariant 4, gated by the esbuild
+ * `platform:'neutral'` purity probe).
+ */
+
+/**
+ * Generate a CSS shorthand for values that come with top, right, bottom, and left
+ *
+ * input: { top: 10, right: 10, bottom: 10, left: 10 }
+ * output: "10px"
+ *
+ * input: { top: 10, right: 20, bottom: 10, left: 20 }
+ * output: "10px 20px"
+ *
+ * input: { top: 10, right: 20, bottom: 30, left: 40 }
+ * output: "10px 20px 30px 40px"
+ *
+ * @param values - The values to generate the shorthand for
+ * @returns The generated shorthand
+ */
+export function generateCSSShorthand(
+  values: {
+    top: number;
+    right: number;
+    bottom: number;
+    left: number;
+  },
+  {
+    ignoreZero = true,
+    suffix = "px",
+  }: {
+    /**
+     * If true and all values are 0, return undefined. Defaults to true.
+     */
+    ignoreZero?: boolean;
+    /**
+     * The suffix to add to the shorthand. Defaults to "px".
+     */
+    suffix?: string;
+  } = {},
+) {
+  const { top, right, bottom, left } = values;
+  if (ignoreZero && top === 0 && right === 0 && bottom === 0 && left === 0) {
+    return undefined;
+  }
+  if (top === right && right === bottom && bottom === left) {
+    return `${top}${suffix}`;
+  }
+  if (right === left) {
+    if (top === bottom) {
+      return `${top}${suffix} ${right}${suffix}`;
+    }
+    return `${top}${suffix} ${right}${suffix} ${bottom}${suffix}`;
+  }
+  return `${top}${suffix} ${right}${suffix} ${bottom}${suffix} ${left}${suffix}`;
+}
+
+/**
+ * Check if an element is visible
+ * @param element - The item to check
+ * @returns True if the item is visible, false otherwise
+ */
+export function isVisible(element: { visible?: boolean }): boolean {
+  return element.visible ?? true;
+}
+
+/**
+ * Rounds a number to two decimal places, suitable for pixel value processing.
+ * @param num The number to be rounded.
+ * @returns The rounded number with two decimal places.
+ * @throws TypeError If the input is not a valid number
+ */
+export function pixelRound(num: number): number {
+  if (isNaN(num)) {
+    throw new TypeError(`Input must be a valid number`);
+  }
+  return Number(Number(num).toFixed(2));
+}
+
+/**
+ * Compile-time exhaustiveness guard for discriminated unions.
+ *
+ * Place in the default branch of a switch over a union: the `value: never` parameter
+ * forces TS to error here if any union member was missed, and the `never` return type
+ * tells control-flow analysis that execution doesn't continue (so callers don't need a
+ * trailing return). Throws at runtime as a defense against type-system bypasses
+ * (`as`, JSON inputs, etc.) — should never actually fire in well-typed code.
+ */
+export function exhaustiveCheck(value: never): never {
+  throw new Error(`Unhandled discriminant: ${String(value)}`);
+}
+
+/**
+ * Serialize a value to JSON with sorted object keys so two equal-but-
+ * differently-ordered objects produce the same string. Used for cache keys
+ * and deep-equality checks where property order isn't a stable guarantee
+ * (e.g. partial TypeStyle entries from Figma's styleOverrideTable).
+ */
+export function stableStringify(value: unknown): string {
+  return JSON.stringify(value, (_key, v) => {
+    if (v && typeof v === "object" && !Array.isArray(v)) {
+      const sorted: Record<string, unknown> = {};
+      for (const k of Object.keys(v as Record<string, unknown>).sort()) {
+        sorted[k] = (v as Record<string, unknown>)[k];
+      }
+      return sorted;
+    }
+    return v;
+  });
+}
