--- src/utils/common.ts
@@ -1,237 +1,0 @@
-import fs from "fs";
-import path from "path";
-import { tagError } from "~/utils/error-meta.js";
-import { isWithin } from "~/utils/local-path.js";
-
-/**
- * Download Figma image and save it locally
- * @param fileName - The filename to save as
- * @param localPath - The local path to save to
- * @param imageUrl - Image URL (images[nodeId])
- * @returns A Promise that resolves to the full file path where the image was saved
- * @throws Error if download fails
- */
-export async function downloadFigmaImage(
-  fileName: string,
-  localPath: string,
-  imageUrl: string,
-): Promise<string> {
-  try {
-    // Ensure local path exists
-    if (!fs.existsSync(localPath)) {
-      fs.mkdirSync(localPath, { recursive: true });
-    }
-
-    const fullPath = path.resolve(path.join(localPath, fileName));
-    if (!isWithin(localPath, fullPath)) {
-      tagError(new Error(`File path escapes target directory: ${fileName}`), {
-        category: "invalid_input",
-      });
-    }
-
-    // Use fetch to download the image
-    const response = await fetch(imageUrl, {
-      method: "GET",
-    });
-
-    if (!response.ok) {
-      tagError(new Error(`Failed to download image: ${response.statusText}`), {
-        category: "image_download",
-      });
-    }
-
-    // Create write stream
-    const writer = fs.createWriteStream(fullPath);
-
-    // Get the response as a readable stream and pipe it to the file
-    const reader = response.body?.getReader();
-    if (!reader) {
-      tagError(new Error("Failed to get response body"), { category: "image_download" });
-    }
-
-    return new Promise((resolve, reject) => {
-      // Process stream
-      const processStream = async () => {
-        try {
-          while (true) {
-            const { done, value } = await reader.read();
-            if (done) {
-              writer.end();
-              break;
-            }
-            writer.write(value);
-          }
-        } catch (err) {
-          writer.end();
-          fs.unlink(fullPath, () => {});
-          reject(err);
-        }
-      };
-
-      // Resolve only when the stream is fully written
-      writer.on("finish", () => {
-        resolve(fullPath);
-      });
-
-      writer.on("error", (err) => {
-        reader.cancel();
-        fs.unlink(fullPath, () => {});
-        reject(new Error(`Failed to write image: ${err.message}`));
-      });
-
-      processStream();
-    });
-  } catch (error) {
-    const errorMessage = error instanceof Error ? error.message : String(error);
-    throw new Error(`Error downloading image: ${errorMessage}`, { cause: error });
-  }
-}
-
-/**
- * Remove keys with empty arrays or empty objects from an object.
- * @param input - The input object or value.
- * @returns The processed object or the original value.
- */
-export function removeEmptyKeys<T>(input: T): T {
-  // If not an object type or null, return directly
-  if (typeof input !== "object" || input === null) {
-    return input;
-  }
-
-  // Handle array type
-  if (Array.isArray(input)) {
-    return input.map((item) => removeEmptyKeys(item)) as T;
-  }
-
-  // Handle object type
-  const result = {} as T;
-  for (const key in input) {
-    if (Object.prototype.hasOwnProperty.call(input, key)) {
-      const value = input[key];
-
-      // Recursively process nested objects
-      const cleanedValue = removeEmptyKeys(value);
-
-      // Skip empty arrays and empty objects
-      if (
-        cleanedValue !== undefined &&
-        !(Array.isArray(cleanedValue) && cleanedValue.length === 0) &&
-        !(
-          typeof cleanedValue === "object" &&
-          cleanedValue !== null &&
-          Object.keys(cleanedValue).length === 0
-        )
-      ) {
-        result[key] = cleanedValue;
-      }
-    }
-  }
-
-  return result;
-}
-
-/**
- * Generate a CSS shorthand for values that come with top, right, bottom, and left
- *
- * input: { top: 10, right: 10, bottom: 10, left: 10 }
- * output: "10px"
- *
- * input: { top: 10, right: 20, bottom: 10, left: 20 }
- * output: "10px 20px"
- *
- * input: { top: 10, right: 20, bottom: 30, left: 40 }
- * output: "10px 20px 30px 40px"
- *
- * @param values - The values to generate the shorthand for
- * @returns The generated shorthand
- */
-export function generateCSSShorthand(
-  values: {
-    top: number;
-    right: number;
-    bottom: number;
-    left: number;
-  },
-  {
-    ignoreZero = true,
-    suffix = "px",
-  }: {
-    /**
-     * If true and all values are 0, return undefined. Defaults to true.
-     */
-    ignoreZero?: boolean;
-    /**
-     * The suffix to add to the shorthand. Defaults to "px".
-     */
-    suffix?: string;
-  } = {},
-) {
-  const { top, right, bottom, left } = values;
-  if (ignoreZero && top === 0 && right === 0 && bottom === 0 && left === 0) {
-    return undefined;
-  }
-  if (top === right && right === bottom && bottom === left) {
-    return `${top}${suffix}`;
-  }
-  if (right === left) {
-    if (top === bottom) {
-      return `${top}${suffix} ${right}${suffix}`;
-    }
-    return `${top}${suffix} ${right}${suffix} ${bottom}${suffix}`;
-  }
-  return `${top}${suffix} ${right}${suffix} ${bottom}${suffix} ${left}${suffix}`;
-}
-
-/**
- * Check if an element is visible
- * @param element - The item to check
- * @returns True if the item is visible, false otherwise
- */
-export function isVisible(element: { visible?: boolean }): boolean {
-  return element.visible ?? true;
-}
-
-/**
- * Rounds a number to two decimal places, suitable for pixel value processing.
- * @param num The number to be rounded.
- * @returns The rounded number with two decimal places.
- * @throws TypeError If the input is not a valid number
- */
-export function pixelRound(num: number): number {
-  if (isNaN(num)) {
-    throw new TypeError(`Input must be a valid number`);
-  }
-  return Number(Number(num).toFixed(2));
-}
-
-/**
- * Compile-time exhaustiveness guard for discriminated unions.
- *
- * Place in the default branch of a switch over a union: the `value: never` parameter
- * forces TS to error here if any union member was missed, and the `never` return type
- * tells control-flow analysis that execution doesn't continue (so callers don't need a
- * trailing return). Throws at runtime as a defense against type-system bypasses
- * (`as`, JSON inputs, etc.) — should never actually fire in well-typed code.
- */
-export function exhaustiveCheck(value: never): never {
-  throw new Error(`Unhandled discriminant: ${String(value)}`);
-}
-
-/**
- * Serialize a value to JSON with sorted object keys so two equal-but-
- * differently-ordered objects produce the same string. Used for cache keys
- * and deep-equality checks where property order isn't a stable guarantee
- * (e.g. partial TypeStyle entries from Figma's styleOverrideTable).
- */
-export function stableStringify(value: unknown): string {
-  return JSON.stringify(value, (_key, v) => {
-    if (v && typeof v === "object" && !Array.isArray(v)) {
-      const sorted: Record<string, unknown> = {};
-      for (const k of Object.keys(v as Record<string, unknown>).sort()) {
-        sorted[k] = (v as Record<string, unknown>)[k];
-      }
-      return sorted;
-    }
-    return v;
-  });
-}
